dataflow/twitter-streaming-pipeline/requirements.txt
google-cloud-pubsub 2.13.11
tweepy 4.12.1
'apache-beam[gcp]'
